+++ src/tui/view/common/tabs.rs
@@ -1,30 +1,28 @@
 use crate::tui::{
     input::Action,
     view::{
-        component::{
-            Component, Draw, DrawContext, Event, Update, UpdateContext,
-        },
-        state::{FixedSelect, StatefulSelect},
+        draw::{Draw, DrawContext},
+        event::{Event, EventHandler, Update, UpdateContext},
+        state::select::{FixedSelect, FixedSelectState},
+        theme::Theme,
     },
 };
-use derive_more::Display;
 use ratatui::prelude::Rect;
 use std::fmt::Debug;
 
 /// Multi-tab display. Generic parameter defines the available tabs.
-#[derive(Debug, Default, Display)]
-#[display(fmt = "Tabs")]
+#[derive(Debug, Default)]
 pub struct Tabs<T: FixedSelect> {
-    tabs: StatefulSelect<T>,
+    tabs: FixedSelectState<T, usize>,
 }
 
 impl<T: FixedSelect> Tabs<T> {
-    pub fn selected(&self) -> T {
+    pub fn selected(&self) -> &T {
         self.tabs.selected()
     }
 }
 
-impl<T: Debug + FixedSelect> Component for Tabs<T> {
+impl<T: FixedSelect> EventHandler for Tabs<T> {
     fn update(&mut self, _context: &mut UpdateContext, event: Event) -> Update {
         match event {
             Event::Input {
@@ -54,7 +52,7 @@ impl<T: FixedSelect> Draw for Tabs<T> {
                 T::iter().map(|e| e.to_string()).collect(),
             )
             .select(self.tabs.selected_index())
-            .highlight_style(context.theme.tab_highlight_style),
+            .highlight_style(Theme::get().tab_highlight_style),
             chunk,
         )
     }