+++ src/tui/view/component/mod.rs
@@ -1,0 +1,10 @@
+//! Specific single-use components
+
+pub mod misc;
+pub mod primary;
+pub mod request;
+pub mod response;
+pub mod root;
+pub mod settings;
+
+pub use root::{FullscreenMode, Root};